Make test reliable on case-insensitive filesystems
I was getting intermittent test failures on my Mac. These
occurred when the test code generated a file path that corresponded
to a real file path on my Mac, but where the case differed. When
this happens the tests fail because the original file path changes
when it's written to JSON due to a call of `getAbsolutePath`
during the write operation.